+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `getQuarter` utility function
+- `addQuarters` utility function
+
 ## [2.3.0] - 2026-04-07
 
 ### Added